Fix critical bug introduced about last week
Re-enable calling callback-function after sending a write-action over the knx-bus.
If we don’t call it, cronjob-tasks won’t be marked as ’ended’ and a cronjob will ’run’ forever.

Though it less than ideal to use EIBd’s callback for this, but at least we don’t call it with an err-object, so we
should be fine.